cloud_backend/requirements.txt
fastapi
uvicorn[standard]
google-genai
google-adk
python-dotenv
websockets
